treat all warning as error

+++ 0x1E-search_algorithms/0-linear.c
@@ -14,7 +14,7 @@ int linear_search(int *array, size_t size, int value)
 {
 	int i = 0;
 
-	for (; i < size; i++)
+	for (; i < (int)size; i++)
 	{
 		printf("Value checked array[%d] = [%d]\n", i, array[i]);
